test(carton): split wrapper normalization coverage
